chore: rename packages in TCK tests to align with SQL extension structure
