+++ jnosql-jakarta-persistence/jnosql-jakarta-persistence-data-tck-runner/src/test/java/org/eclipse/jnosql/extensions/sql/tck/JNoSqlSignatureTests.java
@@ -12,7 +12,7 @@
  *
  *   Ondro Mihalyi
  */
-package org.eclipse.jnosql.tck.jakartapersistence;
+package org.eclipse.jnosql.extensions.sql.tck;
 
 import ee.jakarta.tck.data.standalone.signature.SignatureTests;
 import org.junit.jupiter.api.Disabled;
