+++ jnosql-jakarta-persistence/jnosql-jakarta-persistence-data-tck-runner/src/test/java/org/eclipse/jnosql/extensions/sql/tck/junit/RunOnlyCondition.java
@@ -13,7 +13,7 @@
  *
  *  Ondro Mihalyi
  */
-package org.eclipse.jnosql.tck.jakartapersistence.junit;
+package org.eclipse.jnosql.extensions.sql.tck.junit;
 
 import java.lang.reflect.Method;
 import java.util.Optional;
